+++ oshdb-util/src/main/java/org/heigit/ohsome/oshdb/util/exceptions/OSHDBInvalidTimestampException.java
@@ -3,7 +3,7 @@ package org.heigit.ohsome.oshdb.util.exceptions;
 /**
  * An exception caused by invalid time string input.
  */
-public class OSHDBInvalidTimestampException extends RuntimeException {
+public class OSHDBInvalidTimestampException extends OSHDBException {
   public OSHDBInvalidTimestampException(String message) {
     super(message);
   }